+++ tests_pytest/pytorch/core/mixed_precision/test_resource_utilization.py
@@ -17,7 +17,7 @@ from torch import nn
 
 from model_compression_toolkit.core.pytorch.utils import to_torch_tensor
 from tests_pytest.base_test_classes.base_test_ru_integration import BaseRUIntegrationTester
-from tests_pytest.pytorch.torch_test_mixin import TorchFwMixin
+from tests_pytest.pytorch.torch_test_util.torch_test_mixin import TorchFwMixin
 
 
 class TestRUIntegrationTorch(BaseRUIntegrationTester, TorchFwMixin):
